* ExtensionManager rejects an extension if one of the same name is already loaded or if shutdown() is missing.  * ExtensionManagerDialog won't overwrite a file that's already there.

+++ Jokosher/ExtensionManagerDialog.py
@@ -88,11 +88,19 @@ class ExtensionManagerDialog:
 			if install_response == gtk.RESPONSE_YES:
 				filename = chooser.get_filename()
 				dest = os.path.expanduser("~/.jokosher/extensions/")+os.path.basename(filename)
-				shutil.copy(filename, dest)
-				
-				self.parent.extensionManager.LoadExtensionFromFile(os.path.basename(dest), os.path.dirname(dest))
+				# don't overwite if already there
+				if os.path.exists(dest):
+					messages = ["Filename " + dest + " already exists"]
+					Globals.debug(messages[0])
+				else:
+					# copy the extension and try to load it
+					shutil.copy(filename, dest)
+					messages = self.parent.extensionManager.LoadExtensionFromFile(os.path.basename(dest), os.path.dirname(dest))
+					# if somthing went wrong with the load then revert the copy
+					if len(messages) > 0:
+						os.remove(dest)
+				# TODO: somehow display the failing messages back to the user
 				self.UpdateModel()
-				
 				chooser.destroy()
 			
 			else:
